services/doc_extract/app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
requests==2.32.3

pdfplumber==0.11.4
pymupdf==1.24.10
pytesseract==0.3.13
pillow==10.4.0
opencv-python-headless==4.10.0.84
sqlalchemy==2.0.35
psycopg2-binary==2.9.9